+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGELOG 
 
-## [1.0.2] - 2026-02-05
+## [1.0.3] - 2026-02-19
 
 - Init Idlab Loggable library
 - Add Proxies as default disallowed namespace merged to config
-
+- Fix TreeBuilder for array of strings in idlab_loggable configuration file
+- Add tests for ignored namespaces and classes